generate markets.json from config in token_data.py and market_utils.py
