+++ README.md
@@ -12,23 +12,6 @@ Known token addresses by destination chain:
 
 
 ## Update instructions
-* [updating markets.json](updating_markets.md)
+* [Update instructions](updating.md)
 
 
-## Running the tests
-```
-npm test
-```
-You might need to run `npm install` once before this.
-
-
-## Chain ids
-This repo uses the Wormhole chain ids to reference the chains. For reference, these are:
-
-```python
-Solana = 1
-Ethereum = 2
-Terra = 3
-Binance Smart Chain = 4
-Polygon = 5
-```
