+++ src/config/market_urls.py
@@ -1,0 +1,82 @@
+MARKETS = {
+  "saber": {
+    "name": "Saber",
+    "link": "https://www.saber.so/"
+  },
+  "mercurial": {
+    "name": "Mercurial",
+    "link": "https://mercurial.finance/"
+  },
+  "jupiter": {
+    "name": "Jupiter",
+    "link": "https://jup.ag/"
+  },
+  "atrix": {
+    "name": "Atrix",
+    "link": "https://app.atrix.finance/"
+  },
+  "raydium": {
+    "name": "Raydium",
+    "link": "https://raydium.io/swap/"
+  },
+  "tulip": {
+    "name": "Tulip",
+    "link": "https://tulip.garden/"
+  },
+  "almond": {
+    "name": "Almond",
+    "link": "https://almond.so/"
+  },
+  "orca": {
+    "name": "Orca",
+    "link": "https://www.orca.so/"
+  },
+  "curve": {
+    "name": "Curve",
+    "link": "https://curve.fi/"
+  },
+  "uniswap": {
+    "name": "Uniswap",
+    "link": "https://app.uniswap.org/"
+  },
+  "anchor": {
+    "name": "Anchor",
+    "link": "https://app.anchorprotocol.com/"
+  },
+  "loop": {
+    "name": "Loop Markets",
+    "link": "https://dex.loop.markets/"
+  },
+  "quickswap": {
+    "name": "Quickswap",
+    "link": "https://quickswap.exchange/"
+  },
+  "sushi": {
+    "name": "Sushi",
+    "link": "https://app.sushi.com/"
+  },
+  "pancakeswap": {
+    "name": "Pancakeswap",
+    "link": "https://pancakeswap.finance/swap?inputCurrency=0xfa54ff1a158b5189ebba6ae130ced6bbd3aea76e"
+  },
+  "sunny": {
+    "name": "Sunny",
+    "link": "https://app.sunny.ag/"
+  },
+  "port": {
+    "name": "Port Finance",
+    "link": "https://mainnet.port.finance/#/markets"
+  },
+  "francium": {
+    "name": "Francium",
+    "link": "https://francium.io/app/invest/farm"
+  },
+  "aldrin": {
+    "name": "Aldrin",
+    "link": "https://dex.aldrin.com/pools"
+  },
+  "apricot": {
+    "name": "Apricot",
+    "link": "https://app.apricot.one/"
+  }
+}
