+++ src/dest_ethereum.md
@@ -19,5 +19,5 @@ Known tokens (wormholed to Ethereum)
 | USDCso   | [USD Coin (Wormhole from Solana)](http://coingecko.com/en/coins/usd-coin)           | [0x41f7B8b9b897276b7AAE926a9016935280b44E97](https://etherscan.io/token/0x41f7B8b9b897276b7AAE926a9016935280b44E97)     | solana    | [EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v](https://solscan.io/address/EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v)  | USDCso           |
 | USDTbs   | [Tether USD (Wormhole from BSC)](http://coingecko.com/en/coins/tether)              | [0xDe60aDfDdAAbaAAC3dAFa57B26AcC91Cb63728c4](https://etherscan.io/token/0xDe60aDfDdAAbaAAC3dAFa57B26AcC91Cb63728c4)     | bsc       | [0x55d398326f99059fF775485246999027B3197955](https://bscscan.com/address/0x55d398326f99059fF775485246999027B3197955)     | USDTbs           |
 | USDTso   | [Tether USD (Wormhole from Solana)](http://coingecko.com/en/coins/tether)           | [0x1CDD2EaB61112697626F7b4bB0e23Da4FeBF7B7C](https://etherscan.io/token/0x1CDD2EaB61112697626F7b4bB0e23Da4FeBF7B7C)     | solana    | [Es9vMFrzaCERmJfrF4H2FYD4KCoNkY11McCe8BenwNYB](https://solscan.io/address/Es9vMFrzaCERmJfrF4H2FYD4KCoNkY11McCe8BenwNYB)  | USDTso           |
-| UST      | [UST (Wormhole)](http://coingecko.com/en/coins/terra-usd)                           | [0xa693b19d2931d498c5b318df961919bb4aee87a5](https://etherscan.io/token/0xa693b19d2931d498c5b318df961919bb4aee87a5)     | terra     | [uusd](https://finder.terra.money/columbus-5/address/uusd)                                                               | UST              |
+| UST      | [UST (Wormhole)](http://coingecko.com/en/coins/terra-usd)                           | [0xa693B19d2931d498c5B318dF961919BB4aee87a5](https://etherscan.io/token/0xa693B19d2931d498c5B318dF961919BB4aee87a5)     | terra     | [uusd](https://finder.terra.money/columbus-5/address/uusd)                                                               | UST              |
 | mSOL     | [Marinade staked SOL (Wormhole)](http://coingecko.com/en/coins/marinade-staked-sol) | [0x756bFb452cFE36A5Bc82e4F5f4261A89a18c242b](https://etherscan.io/token/0x756bFb452cFE36A5Bc82e4F5f4261A89a18c242b)     | solana    | [mSoLzYCxHdYgdzU16g5QSh3i5K3z3KZK7ytfqcJm7So](https://solscan.io/address/mSoLzYCxHdYgdzU16g5QSh3i5K3z3KZK7ytfqcJm7So)    | mSOL             |
